server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
requests>=2.31
openai>=1.30
